+++ examples/simulation_multilayer_monochromator_efficiency.py
@@ -56,8 +56,12 @@ eff = pd.DataFrame({
 # calculated by the RAY-UI
 sim.efficiency = eff
 
-#uncomment to run the simulations
-sim.run(multiprocessing=5, force=True, remove_rawrays=True, remove_round_folders=True)
+sim.run(
+    multiprocessing="auto",
+    force=True,
+    remove_rawrays=True,
+    remove_round_folders=True,
+)