+++ examples/simulation_waveFiles.py
@@ -53,7 +53,4 @@ sim.exports  =  [{beamline.Undulator:['RawRaysOutgoing']},
                 {beamline.DetectorAtFocus:['RawRaysOutgoing']}
                 ]
 
-#uncomment to run the simulations
-sim.run(multiprocessing=5, force=True)
-
-
+sim.run(multiprocessing="auto", force=True)